+++ src/modules/auth/auth.service.ts
@@ -2,18 +2,21 @@ import * as bcrypt from 'bcrypt';
 import { Response } from 'express';
 import { ConflictException, Injectable } from '@nestjs/common';
 import { CreateUserDto } from '../user/dto/create-user.dto';
-import { UserService } from '../user/user.service';
 import { AuthTokenService } from './auth-token/auth-token.service';
+import { UserRepository } from '../user/user.repository';
 import { TokenType } from './auth-token/auth-token.types';
 import { AuthTokenPayloadDto } from './auth-token/dto/auth-token-payload.dto';
 import { AUTH_COOKIE_KEY, AUTH_COOKIE_OPTIONS } from './auth.const';
 import { AuthToken } from './auth.types';
+import { ConfigService } from '@nestjs/config';
+import { envKeys } from '../../config/env.const';
 
 @Injectable()
 export class AuthService {
 	constructor(
-		private readonly userService: UserService,
+		private readonly configService: ConfigService,
 		private readonly authTokenService: AuthTokenService,
+		private readonly userRepository: UserRepository,
 	) {}
 
 	/**
@@ -21,7 +24,7 @@ export class AuthService {
 	 * @param createUserDto 이메일, 비밀번호, 닉네임
 	 */
 	async register(createUserDto: CreateUserDto) {
-		const alreadySignedUp = await this.userService.findOne({
+		const alreadySignedUp = await this.userRepository.findOne({
 			email: createUserDto.email,
 		});
 
@@ -29,7 +32,15 @@ export class AuthService {
 			throw new ConflictException('이미 가입된 이메일입니다.');
 		}
 
-		const user = await this.userService.create(createUserDto);
+		const hashedPassword: string = await bcrypt.hash(
+			createUserDto.password,
+			this.configService.get<number>(envKeys.SECURITY.HASH_ROUNDS)!,
+		);
+		const user = await this.userRepository.create({
+			...createUserDto,
+			password: hashedPassword,
+		});
+
 		return this.generateToken({ userId: user._id });
 	}
 
@@ -39,7 +50,10 @@ export class AuthService {
 	 * @param password 비밀번호
 	 */
 	async validateUser(email: string, password: string) {
-		const user = await this.userService.findOne({ email }, { password: 1 });
+		const user = await this.userRepository.findOne(
+			{ email },
+			{ password: 1 },
+		);
 
 		return !!user && (await bcrypt.compare(password, user.password))
 			? user
@@ -51,7 +65,7 @@ export class AuthService {
 	 * @param id 해당 사용자의 아이디
 	 */
 	async withdraw(id: string) {
-		await this.userService.delete(id);
+		await this.userRepository.delete(id);
 	}
 
 	/**
